requirements.txt
atari-py==0.2.9
gym==0.18.3
numpy==1.20.3
Pillow==8.2.0